+++ src/main/java/roomescape/exception/ExceptionApiController.java
@@ -2,16 +2,13 @@ package roomescape.exception;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
-
 import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import roomescape.client.PaymentException;
+import roomescape.payment.PaymentException;
 
 @RestControllerAdvice(annotations = RestController.class)
 public class ExceptionApiController {
